% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRD4_apo}
\alias{BRD4_apo}
\title{This is data to be included in my package}
\description{
A small HDX-MS dataset for BRD4 in apo state
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
