% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRD4_ibet}
\alias{BRD4_ibet}
\title{This is data to be included in my package}
\description{
A complete HDX-MS dataset for BRD4 in complex with IBET151
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
