% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{cleanHDX}
\alias{cleanHDX}
\title{Function that checks a hdx-ms dataset for correct formatting and will
clean up the dataset if specified}
\usage{
cleanHDX(res, clean = TRUE)
}
\arguments{
\item{res}{An object of class DataFrame containing the results of the
hdx-ms  experiment}

\item{clean}{If TRUE will filter multiple charge states and remove missing
values}
}
\value{
Returns a cleaned DataFrame
}
\description{
Function that checks a hdx-ms dataset for correct formatting and will
clean up the dataset if specified
}
\examples{
data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

cleanHDX(res = BRD4_apo, clean = TRUE)



}
