% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{cleanLong}
\alias{cleanLong}
\title{Helper function to clean up long vectors}
\usage{
cleanLong(currentlong, sub, currentbp)
}
\arguments{
\item{currentlong}{numeric vector of current parameter values}

\item{sub}{numeric vector of indices to subset}

\item{currentbp}{numeric vector of current breakpoints}
}
\value{
numeric vector of cleaned up parameter values
}
\description{
Helper function to clean up long vectors
}
\examples{

currentlong <- c(1, 2, NA, 4, 5)
sub <- c(1, 2, 4, 5)
currentbp <- c(1, 3)
cleanLong(currentlong, sub, currentbp)

}
