% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{coverageHeatmap}
\alias{coverageHeatmap}
\title{Plot a coverage heatmap}
\usage{
coverageHeatmap(res, plot = TRUE)
}
\arguments{
\item{res}{An object of class DataFrame containing the results of the
hdx-ms  experiment}

\item{plot}{A logical value indicating whether to plot the heatmap}
}
\value{
Returns a matrix of coverage
}
\description{
Plot a coverage heatmap
}
\examples{
data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

coverageHeatmap(res = BRD4_apo)

}
