% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{doublelogisticFunction}
\alias{doublelogisticFunction}
\title{Implementation of double logistic function (a mixture of logistics)}
\usage{
doublelogisticFunction(t, b, a = 1, q = 1, pi, d)
}
\arguments{
\item{t}{numeric vector of time points}

\item{b}{numeric vector of rate parameters}

\item{a}{numeric vector of plateau parameters}

\item{q}{numeric vector of stretch parameters}

\item{pi}{numeric vector of mixing parameters}

\item{d}{numeric vector of rate parameters}
}
\value{
numeric vector of model predictions
}
\description{
Implementation of double logistic function (a mixture of logistics)
}
\examples{

t <- seq(0, 10, length.out = 100)
b <- c(0.1)
a <- c(1)
q <- c(1)
pi <- c(0.5)
d <- c(0.1)
doublelogisticFunction(t, b, a, q, pi, d)

}
