% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-differential.R
\name{EFDR}
\alias{EFDR}
\title{Function to calculate the corresponding expected False discovery rate
(eFDR) for a given threshold}
\usage{
EFDR(prob, threshold = 0.9)
}
\arguments{
\item{prob}{A numeric vector of probabilities}

\item{threshold}{A numeric value between 0 and 1}
}
\value{
A numeric value indicating the expected FDR
}
\description{
Function to calculate the corresponding expected False discovery rate
(eFDR) for a given threshold
}
\examples{
EFDR(c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1), 0.9)

}
