% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-error.R
\name{error_prediction}
\alias{error_prediction}
\title{Function that computes errors between the predicted peptide values and
the observed peptide values}
\usage{
error_prediction(res, blong, pilong, qlong, dlong, phi)
}
\arguments{
\item{res}{An object of class DataFrame containing the results of the
hdx-ms  experiment}

\item{blong}{A matrix of dimensions R by 1 containing the blong values}

\item{pilong}{A matrix of dimensions R by 1 containing the pilong values}

\item{qlong}{A matrix of dimensions R by 1 containing the qlong values}

\item{dlong}{A matrix of dimensions R by 1 containing the dlong values}

\item{phi}{A numeric value containing the phi value}
}
\value{
Returns exchange indexes as a list
}
\description{
Function that computes errors between the predicted peptide values and
the observed peptide values
}
