% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-helpers.R
\name{initialiser}
\alias{initialiser}
\title{Initialiser}
\usage{
initialiser(uptake_guess, timepoints, numRep, numtimepoints, R, phi)
}
\arguments{
\item{uptake_guess}{A matrix of uptake data}

\item{timepoints}{A vector of timepoints}

\item{numRep}{The number of replicates}

\item{numtimepoints}{The number of timepoints}

\item{R}{The number of residues}

\item{phi}{The maximum uptake possible. Default is 1.}
}
\value{
A matrix of initial parameters
}
\description{
Function to help find good initial parameters
}
