% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{logisticFunction}
\alias{logisticFunction}
\title{Implementation of stretched logistic function}
\usage{
logisticFunction(t, b, a, q = 1)
}
\arguments{
\item{t}{numeric vector of time points}

\item{b}{numeric vector of rate parameters}

\item{a}{numeric vector of plateau parameters}

\item{q}{numeric vector of stretch parameters}
}
\value{
numeric vector of model predictions
}
\description{
Implementation of stretched logistic function
}
\examples{

t <- seq(0, 10, length.out = 100)
b <- c(0.1)
a <- c(1)
q <- c(1)
logisticFunction(t, b, a, q)

}
