% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{loglik}
\alias{loglik}
\title{loglikelihood function for current state of parameters}
\usage{
loglik(
  res,
  pilong,
  blong,
  qlong,
  dlong,
  sigmasq,
  timepoints,
  tCoef = 1,
  index,
  numExch,
  density = "Gaussian",
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{pilong}{numeric vector of mixing parameters}

\item{blong}{numeric vector of rate parameters}

\item{qlong}{numeric vector of stretch parameters}

\item{dlong}{numeric vector of rate parameters}

\item{sigmasq}{numeric variance parameter}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector scaling variance of timepoints}

\item{index}{list of indices for each peptide residues}

\item{numExch}{the number of exchangeable amides for each peptide}

\item{density}{character string of density function.}

\item{phi}{numeric maximum uptake possible}
}
\value{
numeric log likelihood
}
\description{
loglikelihood function for current state of parameters
}
