% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LXRalpha_compounds}
\alias{LXRalpha_compounds}
\title{This is data to be included in my package}
\description{
A list of HDX-mS datasets for LXR-alpha in complexes with 17 ligands.
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
