% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LXRalpha_processed}
\alias{LXRalpha_processed}
\title{This is data to be included in my package}
\description{
LXR-alpha HDX-MS data in the APO state.
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
