% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{marginalEffect}
\alias{marginalEffect}
\title{Function to sample uncertainty of the Uptake}
\usage{
marginalEffect(
  params,
  method = "fitted",
  whichChains = 1,
  tCoef = NULL,
  range = seq.int(4000, 5000)
)
}
\arguments{
\item{params}{An object of class RexParams containing the parameters of
the Rex model}

\item{method}{A character value indicating the method to use.
Either "fitted" or "predict". Default is "fitted" which only include
uncertainty in the model parameters and "predict" includes uncertainty
in the model parameters and the error term (e.g. observation level error)}

\item{whichChains}{An integer value indicating which chain to sample from}

\item{tCoef}{A numeric vector of coefficients to use in the prediction}

\item{range}{A numeric vector of the range of iterations to sample from}
}
\value{
Returns a list of samples
}
\description{
Function to sample uncertainty of the Uptake
}
\examples{
require(RexMS)
require(dplyr)
data("BRD4_apo")
BRD4_apo <- BRD4_apo \%>\% filter(End < 40)

numTimepoints <- length(unique(BRD4_apo$Exposure))
Timepoints <- unique(BRD4_apo$Exposure)

rex_example <- rex(HdxData = DataFrame(BRD4_apo),
                numIter = 4, # typically much larger
                R = max(BRD4_apo$End),
                numtimepoints = numTimepoints,
                timepoints = Timepoints,
                seed = 1L,
                numChains = 1L,
                tCoef = c(0, rep(1, 5)),
                BPPARAM = SerialParam())
rex_example <- RexProcess(HdxData = DataFrame(BRD4_apo),
                          params = rex_example,
                          thin = 1,
                          range = 1:4,
                          whichChains = c(1))
samples <- marginalEffect(rex_example,
                          range= 1:4)                           
}
