% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{maxUptakes}
\alias{maxUptakes}
\title{Function that compuate the maximum uptakes for all sequences in the dataset}
\usage{
maxUptakes(res)
}
\arguments{
\item{res}{An object of class DataFrame containing the results of the
hdx-ms  experiment}
}
\value{
Returns a numeric vector of maximum uptakes
}
\description{
Function that compuate the maximum uptakes for all sequences in the dataset
}
\examples{

data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

maxUptakes(res = BRD4_apo)

}
