% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{metropolisdRate}
\alias{metropolisdRate}
\title{Metropolis step for second rate parameter}
\usage{
metropolisdRate(
  res,
  pilong,
  qlong,
  blong,
  currentd,
  currentSigma,
  bp,
  mean = 0,
  sd = 1,
  dshape = 1,
  drate = 5,
  multivariate = FALSE,
  cormat = NULL,
  timepoints = timepoints,
  tCoef,
  index,
  numExch,
  density,
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{pilong}{numeric vector of mixing parameters}

\item{qlong}{numeric vector of stretch parameters}

\item{blong}{numeric vector of rate parameters}

\item{currentd}{numeric vector of current rate parameters}

\item{currentSigma}{numeric current variance parameter}

\item{bp}{numeric vector of breakpoints}

\item{mean}{numeric mean of the proposal distribution}

\item{sd}{numeric standard deviation of the proposal distribution}

\item{dshape}{numeric shape parameter of the prior gamma distribution}

\item{drate}{numeric rate parameter of the prior gamma distribution}

\item{multivariate}{logical flag for multivariate model}

\item{cormat}{numeric correlation matrix}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector of scaling variance of time points}

\item{index}{list of indices for each peptide residues}

\item{numExch}{list of number of exchangeable amides for each peptide}

\item{density}{character string of density function}

\item{phi}{numeric maximum uptake possible}
}
\value{
numeric updated rate parameters
}
\description{
Metropolis step for second rate parameter
}
