% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{metropolisPi}
\alias{metropolisPi}
\title{metroplis step for the mixing parameter}
\usage{
metropolisPi(
  res,
  currentpi,
  blong,
  qlong,
  dlong,
  currentSigma,
  bp,
  mean = 0,
  sd = 1,
  pishape1 = 1,
  pishape2 = 1,
  multivariate = FALSE,
  cormat = NULL,
  timepoints = timepoints,
  tCoef,
  index,
  numExch,
  density,
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{currentpi}{numeric vector of current mixing parameters}

\item{blong}{numeric vector of rate parameters}

\item{qlong}{numeric vector of stretch parameters}

\item{dlong}{numeric vector of rate parameters}

\item{currentSigma}{numeric current variance parameter}

\item{bp}{numeric vector of breakpoints}

\item{mean}{numeric mean of the proposal distribution}

\item{sd}{numeric standard deviation of the proposal distribution}

\item{pishape1}{numeric shape parameter of the prior beta distribution}

\item{pishape2}{numeric shape parameter of the prior beta distribution}

\item{multivariate}{logical flag for multivariate model}

\item{cormat}{numeric correlation matrix}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector of scaling variance of time points}

\item{index}{list of indices for each peptide residues}

\item{numExch}{list of number of exchangeable amides for each peptide}

\item{density}{character string of density function}

\item{phi}{numeric maximum uptake possible}
}
\value{
numeric updated mixing parameters
}
\description{
metroplis step for the mixing parameter
}
