% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{metropolisRate}
\alias{metropolisRate}
\title{Metropolis step for the rate parameters}
\usage{
metropolisRate(
  res,
  pilong,
  qlong,
  dlong,
  currentb,
  currentSigma,
  bp,
  mean = 0,
  sd = 1,
  shape = 1,
  rate = 5,
  multivariate = FALSE,
  cormat = NULL,
  timepoints = timepoints,
  tCoef,
  index,
  numExch,
  density,
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{pilong}{numeric vector of mixing parameters}

\item{qlong}{numeric vector of stretch parameters}

\item{dlong}{numeric vector of rate parameters}

\item{currentb}{numeric vector of current rate parameters}

\item{currentSigma}{numeric current variance parameter}

\item{bp}{numeric vector of breakpoints}

\item{mean}{numeric mean of the proposal distribution}

\item{sd}{numeric standard deviation of the proposal distribution}

\item{shape}{numeric shape parameter of the prior gamma distribution}

\item{rate}{numeric rate parameter of the prior gamma distribution}

\item{multivariate}{logical flag for multivariate model}

\item{cormat}{numeric correlation matrix}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector of scaling variance of time points}

\item{index}{list of indices for each peptide residues}

\item{numExch}{list of number of exchangeable amides for each peptide}

\item{density}{character string of density function}

\item{phi}{numeric maximum uptake possible}
}
\value{
numeric updated rate parameters
}
\description{
Metropolis step for the rate parameters
}
