% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{metropolisSigma}
\alias{metropolisSigma}
\title{Perform a metropolis step for the variance parameter}
\usage{
metropolisSigma(
  res,
  currentSigma,
  pilong,
  blong,
  qlong,
  dlong,
  mean,
  sd,
  meanprop = 0,
  sdprop = 1,
  multivariate = FALSE,
  cormat = NULL,
  timepoints,
  tCoef,
  index,
  numExch,
  density = "Gaussian",
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{currentSigma}{numeric current variance parameter}

\item{pilong}{numeric vector of mixing parameters}

\item{blong}{numeric vector of rate parameters}

\item{qlong}{numeric vector of stretch parameters}

\item{dlong}{numeric vector of rate parameters}

\item{mean}{numeric mean of the prior distribution}

\item{sd}{numeric standard deviation of the prior distribution}

\item{meanprop}{numeric mean of the proposal distribution}

\item{sdprop}{numeric standard deviation of the proposal distribution}

\item{multivariate}{logical flag for multivariate model}

\item{cormat}{numeric correlation matrix}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector of scaling variance of time points}

\item{index}{list of indices for each peptide residues}

\item{numExch}{list of number of exchangeable amides for each peptide}

\item{density}{character string of density function}

\item{phi}{numeric maximum uptake possible}
}
\value{
numeric updated variance parameter
}
\description{
Perform a metropolis step for the variance parameter
}
