% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_lxr_compound_proccessed}
\alias{out_lxr_compound_proccessed}
\title{This is data to be included in my package}
\description{
The result of running Rex and processing on 17 lxr HDX-MS datasets with
different compounds and differentially comparing to the apo state.
Small data mainly used
for testing and examples and not exemplifying the full capabilities of the
package.
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
