% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_visualisation.R
\name{define_color_function}
\alias{define_color_function}
\alias{hdx_to_pdb_colours}
\alias{view_structure}
\title{Function to set colour ranges for HDX data visualisation onto a pdb
structure}
\usage{
define_color_function(dataset, scale_limits = NULL, cmap_name = NULL)

hdx_to_pdb_colours(
  dataset,
  pdb_filepath,
  scale_limits = NULL,
  cmap_name = NULL
)

view_structure(
  pdb_filepath,
  representation = "cartoon",
  quality = "high",
  color_parameters
)
}
\arguments{
\item{dataset}{A list of estimated protection (positive) and
deprotection (negative) values per residue}

\item{scale_limits}{You can force a range of numerical
values to be mapped only}

\item{cmap_name}{Specifies the name of the color map.
Current options: "ProtDeprot, "viridis", "CSA", "ProbProtDeProt"}

\item{pdb_filepath}{The path to the PDB file that you wish to view}

\item{representation}{A character representing the representation.
Default is "cartoon".}

\item{quality}{The quality of the plot (default is "high").}

\item{color_parameters}{A list of colours and residue numbers to be inputted
into NGLViewer. This is the output of the \code{hdx_to_pdb_colours} function.}
}
\value{
Returns a color function

Returns a list of colours and residue numbers to be inputted into NGLVieweR,
with protection values in Red and deprotection values in Blue, and NA values in Grey.

Returns a structure in the viewer panel
}
\description{
Colour function
}
\examples{

v <- matrix(rnorm(n = 477), nrow = 1)
colnames(v) <- seq.int(ncol(v))

v2 <- v[,seq.int(344, 477), drop = FALSE]
colnames(v2) <- seq.int(ncol(v2))

pdb_filepath <- system.file("extdata", "test_BRD4.pdb", mustWork = TRUE,
 package = "ReX")

mycolor_parameters <- hdx_to_pdb_colours(v2, pdb = pdb_filepath,
 cmap_name = "ProtDeprot")




library(NGLVieweR)
v <- matrix(rnorm(n = 477), nrow = 1)
colnames(v) <- seq.int(ncol(v))

v2 <- v[,seq.int(344, 477), drop = FALSE]
colnames(v2) <- seq.int(ncol(v2))

pdb_filepath <- system.file("extdata", "test_BRD4.pdb", mustWork = TRUE,
 package = "ReX")

mycolor_parameters <- hdx_to_pdb_colours(v2, pdb = pdb_filepath,
 cmap_name = "ProtDeprot")

view_structure(pdb_filepath = pdb_filepath,
 color_parameters = mycolor_parameters)

}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
