% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rex-plotting.R
\name{plotLogLikelihoods}
\alias{plotLogLikelihoods}
\title{plot loglikelihoods across chains}
\usage{
plotLogLikelihoods(rex_params)
}
\arguments{
\item{rex_params}{An object of class RexParams}
}
\value{
Returns a ggplot object
}
\description{
plot loglikelihoods across chains
}
\examples{
require(ReX)
require(dplyr)
require(ggplot2)

data("BRD4_apo")

BRD4_apo <- BRD4_apo \%>\% filter(End < 100)
numTimepoints <- length(unique(BRD4_apo$Exposure))
Timepoints <- unique(BRD4_apo$Exposure)
numPeptides <- length(unique(BRD4_apo$Sequence))

rex_test <- rex(HdxData = DataFrame(BRD4_apo),
             numIter = 10,
             R = max(BRD4_apo$End),
             numtimepoints = numTimepoints,
             timepoints = Timepoints,
             seed = 1L,
             tCoef = c(0, rep(1, 5)),
             BPPARAM = SerialParam())

rex_test <- RexProcess(HdxData = DataFrame(BRD4_apo),
                   params = rex_test,
                   range = 5:10,
                   thin = 1, whichChains = c(1,2))
                   
plotLogLikelihoods(rex_params = rex_test)

}
