% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-conformationalSignatures.R
\name{plotSCSA}
\alias{plotSCSA}
\title{Function to plot the supervised conformational signature analysis results}
\usage{
plotSCSA(
  states.plsda,
  labels,
  whichlabel = 1,
  values = c(brewer.pal(n = 8, name = "PRGn")[c(1, 8)], "grey"),
  type = "catagorical"
)
}
\arguments{
\item{states.plsda}{The OPLS-DA object from the supervisedCSA function}

\item{labels}{The labels to use for the plot. Construct labels carefully
using example below. The labels should be a data frame with the rownames
as the states and the columns as the labels to use for the colouring of the
labels.}

\item{whichlabel}{The column in the labels to use for the analysis. Default
is 1}

\item{values}{The values to use for the colouring of the labels. Default is
\code{brewer.pal(n = 8, name = "PRGn")[c(1,8)]} and "grey" (for unknown labels)}

\item{type}{The type of the labels. Default is "catagorical" but could be
"continuous". If "continuous" the "Unknown" values should be recorded as NAs.}
}
\value{
A ggplot object
}
\description{
Function to plot the supervised conformational signature analysis results
}
\examples{

# Construct labels carefully using known properties of the states (Ligands)
# first a catagorical example

data("out_lxr_compound_proccessed")
data("LXRalpha_compounds")

states <- names(LXRalpha_compounds) 

labels <- data.frame(ABCA1 = rep("Unknown", length(states)),
lipogenic = rep("Unknown", length(states)))
rownames(labels) <- states

labels$ABCA1[rownames(labels) \%in\% c("LXR.623", "AZ9", "AZ8", "AZ5")] <- "low"
labels$ABCA1[rownames(labels) \%in\% c("Az1", "AZ2", "AZ3", "AZ4", "AZ6",
                                    "AZ7", "AZ876", "T0.901317", "WAY.254011",
                                    "F1", "GW3965", "BMS.852927")] <- "high"


labels$lipogenic[rownames(labels) \%in\% c("AZ6", "AZ7", "AZ9",
                                        "AZ8", "GW3965", "BMS.852927",
                                        "LXR.623")] <- "Non-Lipogenic"

labels$lipogenic[rownames(labels) \%in\% c("AZ876", "AZ1",
                                        "T0.901317", "F1", "WAY.254011")] <- "Lipogenic"

labels$ABCA1 <- factor(labels$ABCA1,
                       levels = c("low", "high", "Unknown"))
labels$lipogenic <- factor(labels$lipogenic,
                          levels = c("Non-Lipogenic", "Lipogenic", "Unknown"))

# First using ABCA1 as an example

scsa <- supervisedCSA(RexDifferentialList = out_lxr_compound_proccessed,
                        quantity = "TRE",
                        states = states,
                        labels = labels,
                        whichlabel = "ABCA1",
                        whichTimepoint = 600,
                        orthoI = 1)

plotSCSA(states.plsda = scsa,
        labels = labels,
        whichlabel = "ABCA1")

}
