% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-conformationalSignatures.R
\name{plotTREuncertainty}
\alias{plotTREuncertainty}
\title{Function to plot the uncertainty in the TRE values for different states
e.g. ligands. Typically used to obtain uncertainty in the conformational
landscape.}
\usage{
plotTREuncertainty(df_all, pca_states, states, whichSamples = seq.int(1, 50))
}
\arguments{
\item{df_all}{The data frame containing the TRE values and the residue number
over monte carlo interations. Results from the sampleTREuncertainty function.}

\item{pca_states}{The PCA object from the UnsupervisedCSA function}

\item{states}{The state name to use for the analysis. e.g. ligand used in
differential analysis}

\item{whichSamples}{The samples to use for the analysis.
Default is seq.int(1, 50)}
}
\value{
A ggplot object
}
\description{
Function to plot the uncertainty in the TRE values for different states
e.g. ligands. Typically used to obtain uncertainty in the conformational
landscape.
}
\examples{
library("RexMS")
data(TRE_dist)
data("out_lxr_compound_proccessed")
data("LXRalpha_compounds")

states <- names(LXRalpha_compounds)


ucsa <- UnsupervisedCSA(out_lxr_compound_proccessed,
                       quantity = "TRE",
                       states = states,
                       whichTimepoint = 600,
                       pca_params = list(scale = FALSE,
                                         center = TRUE))
                       
                       
plotTREuncertainty(df_all = TRE_dist,
                  pca_states = ucsa$pca_states,
                  states = states,
                  whichSamples = seq.int(1, 50))

}
