% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{prepareIndexes}
\alias{prepareIndexes}
\title{Function that prepares exchange indexes for a given sequence}
\usage{
prepareIndexes(res)
}
\arguments{
\item{res}{An object of class DataFrame containing the results of the
hdx-ms  experiment}
}
\value{
Returns exchange indexes as a list
}
\description{
Function that prepares exchange indexes for a given sequence
}
\examples{
data("BRD4_apo")
BRD4_apo <- DataFrame(BRD4_apo)

prepareIndexes(res = BRD4_apo)

}
