% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-sampler-utils.R
\name{reversibleJumpBreakPointsAdd}
\alias{reversibleJumpBreakPointsAdd}
\title{Reversible jump step for adding the breakpoints}
\usage{
reversibleJumpBreakPointsAdd(
  res,
  currentSigma,
  currentb,
  currentpi,
  currentbp,
  currentq,
  currentd,
  shape = 1,
  rate = 5,
  logmean = 0,
  logsd = 1,
  shape1 = 1,
  shape2 = 5,
  pishape1 = 1,
  pishape2 = 1,
  dshape = 1,
  drate = 1,
  K,
  R,
  rho,
  lambda,
  multivariate = FALSE,
  cormat = NULL,
  timepoints = timepoints,
  tCoef,
  index,
  numExch,
  density,
  R_lower,
  R_upper,
  phi
)
}
\arguments{
\item{res}{data frame of uptake data}

\item{currentSigma}{numeric current variance parameter}

\item{currentb}{numeric vector of current rate parameters}

\item{currentpi}{numeric vector of current mixing parameters}

\item{currentbp}{numeric vector of current breakpoints}

\item{currentq}{numeric vector of current stretch parameters}

\item{currentd}{numeric vector of current rate parameters}

\item{shape}{numeric shape parameter of the prior gamma distribution}

\item{rate}{numeric rate parameter of the prior gamma distribution}

\item{logmean}{numeric mean of the prior lognormal distribution}

\item{logsd}{numeric standard deviation of the prior lognormal distribution}

\item{shape1}{numeric shape parameter of the prior beta distribution}

\item{shape2}{numeric shape parameter of the prior beta distribution}

\item{pishape1}{numeric shape parameter of the prior beta distribution}

\item{pishape2}{numeric shape parameter of the prior beta distribution}

\item{dshape}{numeric shape parameter of the prior gamma distribution}

\item{drate}{numeric rate parameter of the prior gamma distribution}

\item{K}{numeric number of breakpoints}

\item{R}{numeric maximum value of the breakpoints}

\item{rho}{numeric probability of adding a breakpoint}

\item{lambda}{numeric the mean number of break points}

\item{multivariate}{logical flag for multivariate model}

\item{cormat}{numeric correlation matrix}

\item{timepoints}{numeric vector of time points}

\item{tCoef}{numeric vector of scaling variance of time points}

\item{index}{list of indices for each peptide residues}

\item{numExch}{list of number of exchangeable amides for each peptide}

\item{density}{character string of density function}

\item{R_lower}{numeric lower bound of the breakpoints}

\item{R_upper}{numeric upper bound of the breakpoints}

\item{phi}{numeric maximum uptake possible}
}
\value{
list of updated parameters
}
\description{
Reversible jump step for adding the breakpoints
}
