% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-function.R
\name{RexProcess}
\alias{RexProcess}
\title{proccess Rex results}
\usage{
RexProcess(
  HdxData,
  params,
  thin = 1,
  range = seq.int(4000, 5000, by = thin),
  whichChains = c(1, 2)
)
}
\arguments{
\item{HdxData}{An object of class \code{DataFrame} containing the HDX-MS data}

\item{params}{An object of class \code{RexParams}}

\item{thin}{The thinning interval. Default is 1}

\item{range}{The iterations to keep. Default is seq.int(4000, 5000, by = thin)}

\item{whichChains}{The chains to keep. Default is c(1,2)}
}
\value{
An object of class \code{RexParams} with its summary slot populated
}
\description{
proccess Rex results
}
\examples{
require(ReX)
require(dplyr)
data("BRD4_apo")
BRD4_apo <- BRD4_apo \%>\% filter(End < 40)

numTimepoints <- length(unique(BRD4_apo$Exposure))
Timepoints <- unique(BRD4_apo$Exposure)

rex_example <- rex(HdxData = DataFrame(BRD4_apo),
                numIter = 4, # typically much larger
                R = max(BRD4_apo$End),
                numtimepoints = numTimepoints,
                timepoints = Timepoints,
                seed = 1L,
                numChains = 1L,
                tCoef = c(0, rep(1, 5)),
                BPPARAM = SerialParam())
rex_example <- RexProcess(HdxData = DataFrame(BRD4_apo),
                          params = rex_example,
                          thin = 1,
                          range = 1:4,
                          whichChains = c(1))
}
