% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ReX-methods.R
\docType{class}
\name{RexChains-class}
\alias{RexChains-class}
\alias{.RexChains}
\alias{RexSummary-class}
\alias{.RexSummary}
\alias{RexParams-class}
\alias{.RexParams}
\alias{RexChain-class}
\alias{.RexChain}
\alias{RexDifferential-class}
\alias{.RexDifferential}
\alias{chains}
\alias{show,RexParams-method}
\alias{show,RexChain-method}
\alias{show,RexSummary-method}
\alias{length,RexChains-method}
\alias{length,RexParams-method}
\alias{posteriorEstimates}
\alias{posteriorEstimates,RexSummary-method}
\alias{Rex.quantiles}
\alias{Rex.quantiles,RexSummary-method}
\alias{Rex.resolution}
\alias{Rex.resolution,RexSummary-method}
\alias{Rex.peptideError}
\alias{Rex.peptideError,RexSummary-method}
\alias{Rex.globals}
\alias{Rex.globals,RexSummary-method}
\alias{posteriorEstimates,RexParams-method}
\alias{Rex.quantiles,RexParams-method}
\alias{Rex.resolution,RexParams-method}
\alias{Rex.peptideError,RexParams-method}
\alias{Rex.globals,RexParams-method}
\alias{[[,RexChains,ANY,ANY-method}
\alias{[,RexChains,ANY,ANY,ANY-method}
\alias{[,RexParams,ANY,ANY,ANY-method}
\alias{show,RexChains-method}
\alias{show,RexDifferential-method}
\alias{Rex.predictionError}
\alias{Rex.estimates}
\alias{Rex.probs}
\alias{Rex.eFDR}
\alias{Rex.predictionError,RexDifferential-method}
\alias{Rex.estimates,RexDifferential-method}
\alias{Rex.probs,RexDifferential-method}
\alias{Rex.eFDR,RexDifferential-method}
\title{Infrastructure to to store and process MCMC results}
\usage{
chains(object)

\S4method{show}{RexParams}(object)

\S4method{show}{RexChain}(object)

\S4method{show}{RexSummary}(object)

\S4method{length}{RexChains}(x)

\S4method{length}{RexParams}(x)

posteriorEstimates(object)

\S4method{posteriorEstimates}{RexSummary}(object)

Rex.quantiles(object)

\S4method{Rex.quantiles}{RexSummary}(object)

Rex.resolution(object)

\S4method{Rex.resolution}{RexSummary}(object)

Rex.peptideError(object)

\S4method{Rex.peptideError}{RexSummary}(object)

Rex.globals(object)

\S4method{Rex.globals}{RexSummary}(object)

\S4method{posteriorEstimates}{RexParams}(object)

\S4method{Rex.quantiles}{RexParams}(object)

\S4method{Rex.resolution}{RexParams}(object)

\S4method{Rex.peptideError}{RexParams}(object)

\S4method{Rex.globals}{RexParams}(object)

\S4method{[[}{RexChains,ANY,ANY}(x, i, j = "missing", drop = "missing")

\S4method{[}{RexChains,ANY,ANY,ANY}(x, i, j = "missing", drop = "missing")

\S4method{[}{RexParams,ANY,ANY,ANY}(x, i, j = "missing", drop = "missing")

\S4method{show}{RexChains}(object)

\S4method{show}{RexDifferential}(object)

Rex.predictionError(object)

Rex.estimates(object)

Rex.probs(object)

Rex.eFDR(object)

\S4method{Rex.predictionError}{RexDifferential}(object)

\S4method{Rex.estimates}{RexDifferential}(object)

\S4method{Rex.probs}{RexDifferential}(object)

\S4method{Rex.eFDR}{RexDifferential}(object)
}
\arguments{
\item{object}{An instance of appropriate class.}

\item{x}{Object to be subset.}

\item{i}{An \code{integer()}. Should be of length 1 for \code{[[}.}

\item{j}{Missing.}

\item{drop}{Missing.}
}
\value{
An object of class \code{RexParams} which stores the main results
for the analysis when using Rex
}
\description{
The \code{RexParams} infrastructure is used to store and process MCMC results for
Rex model from Crook et al 2024.
}
\details{
Objects of the \code{RexParams} class are created with the \code{Rex()} function
These objects store the \emph{priors} for the model and the results of the MCMC
chains, which themselves are stored as an instance of class \code{RexChains} and
can be accessed with the \code{chains()} function. A summary of the \code{RexChains}
(or class \code{RexSummary}) can be further computed with the \code{RexProcess}
function.

see the \emph{RexMS} vignette for examples
}
\section{Slots}{

\describe{
\item{\code{chains}}{\code{list()} containing the individual full MCMC chain
results in an \code{RexChains} instance. Each element must be a
valid \code{RexChain} instance.}

\item{\code{posteriorEstimates}}{A \code{DataFrame} documenting the posteriors
in an \code{RexSummary} instance}

\item{\code{diagnostics}}{A \code{matrix} of dimensions 1 by 2 containing the
\code{RexSummary} diagnostics.}

\item{\code{Rex.quantiles}}{A \code{DataFrame} of with R rows indicated quantiles for key
parameters in the Rex model}

\item{\code{Rex.globals}}{A \code{DataFrame} of with R rows indicated global parameters
at the moment the standard deviation sigma and the model likelihoods}

\item{\code{Rex.resolution}}{A \code{DataFrame} of with R rows indicated resolution metrics}

\item{\code{Rex.peptideError}}{A \code{DataFrame} indicated peptide error}

\item{\code{method}}{A \code{character()} storing the Rex method name}

\item{\code{priors}}{A \code{list()} with the priors for the parameters}

\item{\code{seed}}{An \code{integer()} with the random number generation seed.}

\item{\code{summary}}{Object of class \code{RexSummary} the summarised MCMC results
available in the \code{RexParams} instance.}

\item{\code{chains}}{Object of class \code{RexChains} containing the full MCMC results
in the \code{RexParams} instance}

\item{\code{interval}}{A \code{numeric()} with the interval for the residues cover}

\item{\code{dataset}}{\code{character} indicating the dataset}

\item{\code{blong}}{\code{matrix} of dimensions R by numIter containing the
blong values for the residues}

\item{\code{pilong}}{\code{matrix} of dimensions R by numIter containing the
pilong values for the residues}

\item{\code{qlong}}{\code{matrix} of dimensions R by numIter containing the
qlong values for the residues}

\item{\code{dlong}}{\code{matrix} of dimensions R by numIter containing the
dlong values for the residues}

\item{\code{loglikelihood}}{\code{numeric} of length numIter containing the
loglikelihood values}

\item{\code{phi}}{\code{numeric} of length 1 containing the phi values for the model}

\item{\code{UptakeGuess}}{\code{list} each with a matrix of dimensions R by numTimepoints
containing the UptakeGuess values for the residues}

\item{\code{numBreak}}{\code{numeric} of length numIter containing the number of breakpoints}

\item{\code{Sigma}}{\code{numeric} of length numIter containing the Sigma values}

\item{\code{R}}{\code{integer} of length 1 containing the number of residues}

\item{\code{timepoints}}{\code{numeric} of length numTimepoints containing the timepoints}

\item{\code{numIter}}{\code{integer} of length 1 containing the number of mcmc iterations}

\item{\code{numPeptides}}{\code{integer} of length 1 containing the number of peptides}

\item{\code{numTimepoints}}{\code{integer} of length 1 containing the number of timepoints}

\item{\code{dataset}}{\code{character} indicating the dataset}

\item{\code{Rex.predictionError}}{\code{DFrame} containing the prediction error}

\item{\code{Rex.estimates}}{\code{DFrame} containing the estimates ARE, TRE, signed ARE}

\item{\code{Rex.probs}}{\code{DFrame} containing the probabilities and total probabilities}

\item{\code{Rex.eFDR}}{\code{DFrame} containing the eFDR}
}}

