% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-conformationalSignatures.R
\name{sampleTREuncertainty}
\alias{sampleTREuncertainty}
\title{Function to calculate the uncertainty in the TRE values by sampling the
posterior distribution of the Rex parameters. Typically used to obtain
uncertainty in the conformational signatures.}
\usage{
sampleTREuncertainty(
  HdxData,
  RexParamsList,
  quantity = "TRE",
  states,
  whichChain = 1,
  whichSamples = seq.int(1, 50),
  whichTimepoint = 600,
  num_montecarlo = 5000
)
}
\arguments{
\item{HdxData}{The HdxData object}

\item{RexParamsList}{A list of RexParams objects}

\item{quantity}{The quantity to use for the analysis. Default is "TRE"}

\item{states}{The state name to use for the analysis. e.g. ligand used in
differential analysis}

\item{whichChain}{The chain to use for the analysis. Default is 1}

\item{whichSamples}{The samples to use for the analysis.
Default is seq.int(1, 50)}

\item{whichTimepoint}{The timepoint to use for the analysis. Default is 600.}

\item{num_montecarlo}{The number of montecarlo samples to use for the
analysis}
}
\value{
A data frame containing the TRE values and the residue number over
monte carlo interations
}
\description{
Function to calculate the uncertainty in the TRE values by sampling the
posterior distribution of the Rex parameters. Typically used to obtain
uncertainty in the conformational signatures.
}
\examples{

  
\dontrun{
library("RexMS")
data("LXRalpha_processed")
data("rex_lxr")
data("LXRalpha_compounds")

states <- names(LXRalpha_compounds)

TRE_dist <- sampleTREuncertainty(HdxData = LXRalpha_processed,
                               RexParamsList = rex_lxr,
                               quantity = "TRE",
                               states = states,
                               whichChain = 1,
                               whichSamples = seq.int(1, 50),
                               whichTimepoint = 600,
                               num_montecarlo = 5000)
                               


}

}
