% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rex-conformationalSignatures.R
\name{supervisedCSA}
\alias{supervisedCSA}
\title{Function to perform supervised conformational signature analysis. In this
case the construction of the signatures using labels as part of the
dimensionality reduction which are defined in labels. The type of the labels
can be catagorical or continuous. The function returns the OPLS-DA object.
The labels are allowed to contain "Unknown" values which are ignored in the
analysis if the type is catagorical. If the type is continuous, the "Unknown"
values should be recorded as NAs.}
\usage{
supervisedCSA(
  RexDifferentialList,
  quantity = "TRE",
  states,
  labels,
  whichlabel = 1,
  whichTimepoint = 600,
  type = "catagorical",
  orthoI = 1
)
}
\arguments{
\item{RexDifferentialList}{A list of RexDifferential objects}

\item{quantity}{The quantity to use for the analysis. Default is "TRE"}

\item{states}{The state name to use for the analysis. e.g. ligand used in
differential analysis}

\item{labels}{The labels to use for the analysis. Construct labels carefully
using example below. The labels should be a data frame with the rownames
as the states and the columns as the labels to use for the analysis.}

\item{whichlabel}{The column in the labels to use for the analysis. Default
is 1}

\item{whichTimepoint}{The timepoint to use for the analysis. Default is 600}

\item{type}{The type of the labels. Default is "catagorical" but could be
"continuous". If "continuous" the "Unknown" values should be recorded as NAs.}

\item{orthoI}{The number of orthogonal components to use in the OPLS-DA
analysis. Default is 1}
}
\value{
An OPLS-DA object
}
\description{
Function to perform supervised conformational signature analysis. In this
case the construction of the signatures using labels as part of the
dimensionality reduction which are defined in labels. The type of the labels
can be catagorical or continuous. The function returns the OPLS-DA object.
The labels are allowed to contain "Unknown" values which are ignored in the
analysis if the type is catagorical. If the type is continuous, the "Unknown"
values should be recorded as NAs.
}
\examples{

# Construct labels carefully using known properties of the states (Ligands)
# first a catagorical example

data("out_lxr_compound_proccessed")
data("LXRalpha_compounds")

states <- names(LXRalpha_compounds) 

labels <- data.frame(ABCA1 = rep("Unknown", length(states)),
lipogenic = rep("Unknown", length(states)))
rownames(labels) <- states

labels$ABCA1[rownames(labels) \%in\% c("LXR.623", "AZ9", "AZ8", "AZ5")] <- "low"
labels$ABCA1[rownames(labels) \%in\% c("Az1", "AZ2", "AZ3", "AZ4", "AZ6",
                                    "AZ7", "AZ876", "T0.901317", "WAY.254011",
                                    "F1", "GW3965", "BMS.852927")] <- "high"


labels$lipogenic[rownames(labels) \%in\% c("AZ6", "AZ7", "AZ9",
                                        "AZ8", "GW3965", "BMS.852927",
                                        "LXR.623")] <- "Non-Lipogenic"

labels$lipogenic[rownames(labels) \%in\% c("AZ876", "AZ1",
                                        "T0.901317", "F1", "WAY.254011")] <- "Lipogenic"

labels$ABCA1 <- factor(labels$ABCA1,
                       levels = c("low", "high", "Unknown"))
labels$lipogenic <- factor(labels$lipogenic,
                          levels = c("Non-Lipogenic", "Lipogenic", "Unknown"))

# First using ABCA1 as an example

scsa <- supervisedCSA(RexDifferentialList = out_lxr_compound_proccessed,
                        quantity = "TRE",
                        states = states,
                        labels = labels,
                        whichlabel = "ABCA1",
                        whichTimepoint = 600,
                        orthoI = 1)
                        
# Now using lipogenic as an example

scsa2 <- supervisedCSA(RexDifferentialList = out_lxr_compound_proccessed,
                       quantity = "TRE",
                       states = states,
                       labels = labels,
                       whichlabel = "lipogenic",
                       whichTimepoint = 600,
                       orthoI = 1)

# Now using a continuous example, add additional annotation to the labels
# Here we use the ED50 values of the ligands, using the log values to 
# because of the large range

labels$ED50 <- NA

labels[, "ED50"] <- log(c(4.11, NA, NA, 0.956, NA, 9.64, 1.49,
                         5.65, 0.969, 2.10, 11.3, NA, 31.5, 341, 32.2, 17.2,
                          NA))
                          
scsa3 <- supervisedCSA(RexDifferentialList = out_lxr_compound_proccessed,
                       quantity = "TRE",
                       states = states,
                       labels = labels,
                       whichlabel = "ED50",
                       whichTimepoint = 600,
                       type = "continuous",
                       orthoI = 1)             

}
