% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRE_dist}
\alias{TRE_dist}
\title{This is data to be included in my package}
\description{
The results of running the following function:
sampleTREuncertainty(HdxData = LXRalpha_processed,
RexParamsList = rex_lxr,
quantity = "TRE",
states = states,
whichChain = 1,
whichSamples = seq.int(1, 50),
whichTimepoint = 600,
num_montecarlo = 5000)
and storing the output for use as an example for other functions.
}
\author{
Oliver Crook \email{ococrook@gmail.com}
}
\keyword{data}
