% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReX-utils.R
\name{uptakePredict}
\alias{uptakePredict}
\title{Function that predict the uptake from the parameters of a Rex model}
\usage{
uptakePredict(params)
}
\arguments{
\item{params}{An object of class RexParams containing the parameters of the Rex model}
}
\value{
Returns a DataFrame of the predicted uptake
}
\description{
Function that predict the uptake from the parameters of a Rex model
}
\examples{
require(RexMS)
require(dplyr)
data("BRD4_apo")
BRD4_apo <- BRD4_apo \%>\% filter(End < 40)

numTimepoints <- length(unique(BRD4_apo$Exposure))
Timepoints <- unique(BRD4_apo$Exposure)

rex_example <- rex(HdxData = DataFrame(BRD4_apo),
                numIter = 4, # typically much larger
                R = max(BRD4_apo$End),
                numtimepoints = numTimepoints,
                timepoints = Timepoints,
                seed = 1L,
                numChains = 1L,
                tCoef = c(0, rep(1, 5)),
                BPPARAM = SerialParam())
rex_example <- RexProcess(HdxData = DataFrame(BRD4_apo),
                          params = rex_example,
                          thin = 1,
                          range = 1:4,
                          whichChains = c(1))
                          
uptakePredict(rex_example)                                      
}
